# Conflicts:
#	gradle/forge.versions.toml
#	src/main/java/com/gregtechceu/gtceu/api/block/MetaMachineBlock.java
#	src/main/java/com/gregtechceu/gtceu/api/cover/filter/SimpleItemFilter.java
#	src/main/java/com/gregtechceu/gtceu/api/data/worldgen/GTOreDefinition.java
#	src/main/java/com/gregtechceu/gtceu/api/fluids/FluidBuilder.java
#	src/main/java/com/gregtechceu/gtceu/api/machine/MetaMachine.java
#	src/main/java/com/gregtechceu/gtceu/api/machine/trait/customlogic/SteamBoilerLogic.java
#	src/main/java/com/gregtechceu/gtceu/api/pattern/BlockPattern.java
#	src/main/java/com/gregtechceu/gtceu/api/pattern/Predicates.java
#	src/main/java/com/gregtechceu/gtceu/api/recipe/GTRecipeType.java
#	src/main/java/com/gregtechceu/gtceu/api/registry/GTRegistries.java
#	src/main/java/com/gregtechceu/gtceu/api/registry/registrate/provider/GTBlockstateProvider.java
#	src/main/java/com/gregtechceu/gtceu/client/ClientEventListener.java
#	src/main/java/com/gregtechceu/gtceu/client/model/pipe/ActivablePipeModel.java
#	src/main/java/com/gregtechceu/gtceu/client/model/pipe/PipeModel.java
#	src/main/java/com/gregtechceu/gtceu/client/renderer/item/decorator/GTLampItemOverlayRenderer.java
#	src/main/java/com/gregtechceu/gtceu/client/renderer/machine/impl/BoilerMultiPartRender.java
#	src/main/java/com/gregtechceu/gtceu/common/CommonEventListener.java
#	src/main/java/com/gregtechceu/gtceu/common/CommonProxy.java
#	src/main/java/com/gregtechceu/gtceu/common/block/LampBlock.java
#	src/main/java/com/gregtechceu/gtceu/common/block/LaserPipeBlock.java
#	src/main/java/com/gregtechceu/gtceu/common/block/OpticalPipeBlock.java
#	src/main/java/com/gregtechceu/gtceu/common/block/RubberLogBlock.java
#	src/main/java/com/gregtechceu/gtceu/common/blockentity/LaserPipeBlockEntity.java
#	src/main/java/com/gregtechceu/gtceu/common/blockentity/OpticalPipeBlockEntity.java
#	src/main/java/com/gregtechceu/gtceu/common/commands/GTCommands.java
#	src/main/java/com/gregtechceu/gtceu/common/cover/WirelessTransmitterCover.java
#	src/main/java/com/gregtechceu/gtceu/common/data/GTAttributeModifierIds.java
#	src/main/java/com/gregtechceu/gtceu/common/data/GTBlockStateProperties.java
#	src/main/java/com/gregtechceu/gtceu/common/data/GTDimensionMarkers.java
#	src/main/java/com/gregtechceu/gtceu/common/data/GTPlaceholders.java
#	src/main/java/com/gregtechceu/gtceu/common/data/GTRecipeConditions.java
#	src/main/java/com/gregtechceu/gtceu/common/data/item/GTDataComponents.java
#	src/main/java/com/gregtechceu/gtceu/common/data/models/GTModels.java
#	src/main/java/com/gregtechceu/gtceu/common/datafixer/fixes/DataItemComponentFix.java
#	src/main/java/com/gregtechceu/gtceu/common/item/GTBucketItem.java
#	src/main/java/com/gregtechceu/gtceu/common/item/behavior/DataItemBehavior.java
#	src/main/java/com/gregtechceu/gtceu/common/item/behavior/PortableScannerBehavior.java
#	src/main/java/com/gregtechceu/gtceu/common/item/modules/TextModuleBehaviour.java
#	src/main/java/com/gregtechceu/gtceu/common/item/tool/behavior/HarvestCropsBehavior.java
#	src/main/java/com/gregtechceu/gtceu/common/item/tool/behavior/ProspectingBehavior.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/multiblock/electric/monitor/MonitorGroup.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/owner/ArgonautsOwner.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/storage/CreativeChestMachine.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/trait/customlogic/LargeBoilerLogic.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/trait/customlogic/SmallBoilerLogic.java
#	src/main/java/com/gregtechceu/gtceu/common/machine/trait/miner/MinerLogic.java
#	src/main/java/com/gregtechceu/gtceu/common/recipe/condition/HeraclesQuestCondition.java
#	src/main/java/com/gregtechceu/gtceu/core/MixinHelpers.java
#	src/main/java/com/gregtechceu/gtceu/data/recipe/misc/FuelRecipes.java
#	src/main/java/com/gregtechceu/gtceu/integration/emi/oreprocessing/GTOreProcessingEmiCategory.java
#	src/main/java/com/gregtechceu/gtceu/integration/emi/recipe/GTRecipeEMICategory.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/GTKubeJSPlugin.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/block/CoilBlockBuilder.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/machine/IMachineBuilderKJS.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/machine/KJSSteamMachineBuilder.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/machine/KJSTieredMachineBuilder.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/builders/machine/KJSTieredMultiblockBuilder.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/events/GTBedrockFluidVeinEventJS.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/events/GTBedrockOreVeinEventJS.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/events/GTOreVeinEventJS.java
#	src/main/java/com/gregtechceu/gtceu/integration/kjs/recipe/components/GTRecipeComponents.java
#	src/main/java/com/gregtechceu/gtceu/integration/xei/widgets/GTOreVeinWidget.java
#	src/main/resources/gtceu.mixins.json

+++ src/main/java/com/gregtechceu/gtceu/integration/cctweaked/peripherals/CentralMonitorPeripheral.java
@@ -96,7 +96,7 @@ public class CentralMonitorPeripheral implements GenericPeripheral {
         }
 
         @LuaFunction
-        public void setScale(double scale) {
+        public void setScale(float scale) {
             if (getModuleItem() instanceof TextModuleBehaviour textModule) {
                 textModule.setScale(stack, scale);
             }
